+++ frontend/components/checkbox/CheckBox.tsx
@@ -1,7 +1,44 @@
 import { ButtonHTMLAttributes, FC } from "react";
-import { checkBoxCVA } from "./styles.cva";
-import { VariantProps } from "cva";
+import { cva, VariantProps } from "cva";
 import CheckIcon from "@assets/svg/files/check.svg";
+
+export const checkBoxCVA = cva("button", {
+  variants: {
+    rounding: {
+      sm: ["rounded-sm"],
+      md: ["rounded"],
+      lg: ["rounded-lg"],
+      full: ["rounded-full"],
+    },
+    size: {
+      small: ["text-md", "p-1"],
+      medium: ["text-lg", "p-3"],
+      large: ["text-xl", "p-4"],
+    },
+    border: {
+      none: ["border-0"],
+      border: ["border"],
+      underline: ["border-0", "border-b", "!px-0", "!rounded-none", "!pb-0"],
+    },
+    disabled: {
+      true: ["opacity-50 pointer-events-none"],
+      false: "",
+    },
+  },
+  compoundVariants: [
+    {
+      rounding: "md",
+      border: "border",
+      class: "",
+    },
+  ],
+  defaultVariants: {
+    rounding: "md",
+    size: "small",
+    border: "border",
+    disabled: false,
+  },
+});
 
 export interface CheckProps extends ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof checkBoxCVA> {
   checked: boolean;
