+++ frontend/components/checkbox/index.ts
@@ -1,1 +1,2 @@
 export { default as CheckBox } from "./CheckBox";
+export { default as CustomCheck } from "./CustomCheck";
