+++ frontend/pages/contacts/components/ICRC/contactAssetPop.tsx
@@ -3,7 +3,7 @@ import { IconType, IconTypeEnum } from "@/const";
 import { ReactComponent as PlusIcon } from "@assets/svg/files/plus.svg";
 //
 import { CustomButton } from "@components/button";
-import { CustomCheck } from "@components/CheckBox";
+import { CustomCheck } from "@components/checkbox";
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu";
 import { Asset, AssetToAdd } from "@redux/models/AccountModels";
 import { Fragment, useEffect, useState } from "react";
